Add homepage navbar in the login/sign-up page
[finishes #105085536]
The user can click the Resly link in the navbar while in login/sign-up page of either the restaurateur or diner and be redirected to their respective homepage.
While the user is on the login/sign-up page, they can click the sign-up/login link on the navbar and be redirected to the corresponding page.