Add Gitter badge
Gitter is an instant messaging and chat room system for developers
and users of GitHub repositories. Gitter is provided as software-as-a-service,
with a free option providing all basic features and the ability to create
a single private chat room, and paid subscription options for individuals
and organisations, which allows them to create arbitrary numbers of private chat rooms.