[Feature #117808237] Restaurant Setting
A settings menu/link is added to the navigation bar.
When creating a restaurant, a user optionally adds a refund rate. If not provided,
this defaults to 0.
A new migration is created to include a refund_rate column to the restaurant table.
This column is set to fillable in the restaurant model.