Ft booking multiple tables 117327517 (#182)
* [Feature #117327517] Booking Multiple Tables

Users can book multiple tables at a time aside just booking for a table.
Users cannot access a table that has already been booed.
A is_booked column is added to the tables model.
New routes are added to handle jquery requests for booking.
Booking is performed with the Session class.

* [Feature #117327517] Booking Multiple Tables

The migration file for the is_booked table is changed from ...add_booked_col_to_tables... to ...add_is_booked_col_to_tables...

* [Fearture #117327517] Booking Multiple Tables

The cart is restructured to allow for multiple bookings.
Only signed in users have access to the page. Other users are redirected to the login page.
Add more status information about tables including tables on-hold, tables available and tables that have been booked.

* [Hotfix #117327517] Fix issue with down part of the new migration

The down part of the new migration failed because of structure.
This has been restructured to one-line code.

* [Chores] Fix merge conflicts

* [Chores #117327517] Booking multiple tables

The array_map is properly formatted to be more readable.
The table status of is_on_hold is changed to is_selected.
The css for the book is properly aligned.
Irrelevant comments are removed from the code.

* [Chores #117327517] Add test for multiple and single booking

Tests are added to ensure expected functionalities for single booking
and multiple booking of tables.

* [Chores #117327517] Add test for the booking page

Tests are added to ensure expected functionalities for single booking
and multiple booking of tables.