Ft restaurant settings 117808237 (#185)
* [Feature #117808237] Restaurant Setting

A settings menu/link is added to the navigation bar.
When creating a restaurant, a user optionally adds a refund rate. If not provided,
this defaults to 0.
A new migration is created to include a refund_rate column to the restaurant table.
This column is set to fillable in the restaurant model.

* [Feature #117808237] Modify the refund rate as a restaurant property

The refund rate is modified to belong to the restaurant model as a property.
When a user cancels a reservation, the refund is calculated using the
rate that is set by the owner of the restaurant.

* [Chores #117808237] Fix the failing tests and merge conflicts

Some of the tests that failed are fixed together with the merge conflicts.

* [Chores #117808237] Remove irrelevant code

Some irrelevant code present in the application is removed.