[HotFix #119262913] Fix the bug with the cancelled reservations (#186)
* [HotFix #119262913] Fix the bug with the cancelled reservations

The list of cancelled reservations contains results including
those that belong to other users. This is fixed.
Once a reservation is cancelled, the table is released.
This means the table is not selected or booked

Finishes #119262913.

* [Fix #119262913] Fix failing tests

The previous ratings model is removed as it is no longer needed.
The current ratings models is updated with a down method.